move utils.h
