+++ ocos/utils.h
